Added ImageManipConfig message
