+++ include/depthai-shared/datatype/DatatypeEnum.hpp
@@ -4,7 +4,7 @@
 
 namespace dai {
 
-enum class DatatypeEnum : std::int32_t { Buffer, ImgFrame, NNData };
+enum class DatatypeEnum : std::int32_t { Buffer, ImgFrame, NNData, ImageManipConfig };
 bool isDatatypeSubclassOf(DatatypeEnum parent, DatatypeEnum children);
 
 }  // namespace dai
