+++ src/datatype/DatatypeEnum.cpp
@@ -21,7 +21,10 @@ struct hash<dai::DatatypeEnum> {
 namespace dai {
 
 const std::unordered_map<DatatypeEnum, std::vector<DatatypeEnum>> hierarchy = {
-    {DatatypeEnum::Buffer, {DatatypeEnum::ImgFrame, DatatypeEnum::NNData}}, {DatatypeEnum::ImgFrame, {}}, {DatatypeEnum::NNData, {}}};
+    {DatatypeEnum::Buffer, {DatatypeEnum::ImgFrame, DatatypeEnum::NNData, DatatypeEnum::ImageManipConfig}},
+    {DatatypeEnum::ImgFrame, {}},
+    {DatatypeEnum::NNData, {}},
+    {DatatypeEnum::ImageManipConfig, {}}};
 
 bool isDatatypeSubclassOf(DatatypeEnum parent, DatatypeEnum children) {
     for(const auto& d : hierarchy.at(parent)) {
